Upgrade to version 22 of nextcloud
This is also a refactor of the way we where doing the install. This
branches is made to improve composability by adding the "well-known" (at
least in our organisation) `docker-entrypoint.d` folder. I rename some
of the env var to match the "official" docker image and make the
transition from the official image to this image smoother.

I didn't test this image against openshift (I'm thinking about running
as non-root) please let me know if you find some incompatibilities.